polish(zlx): move bitflip out of classic_control
