+++ dizoo/bitflip/envs/bitflip_env.py
@@ -1,7 +1,6 @@
 import copy
 import random
 import numpy as np
-
 from typing import Any, Dict, Optional, Union, List
 
 from ding.envs import BaseEnv, BaseEnvInfo, BaseEnvTimestep
